add new conf param

+++ shell_tests/run_tests.py
@@ -243,7 +243,8 @@ class TestsRunner(object):
             else:
                 report = self.run_tests(cs_handler)
             finally:
-                self.delete_cloudshell_on_do()
+                if not self.conf.not_delete_cs:
+                    self.delete_cloudshell_on_do()
 
         if not attempts and error:
             raise error
